+++ stroom-core-shared/src/main/java/stroom/streamtask/shared/FindStreamTaskCriteria.java
@@ -35,16 +35,12 @@ import stroom.streamstore.shared.StreamType;
  * </p>
  */
 public final class FindStreamTaskCriteria extends BaseCriteria implements HasIsConstrained {
-    public static final int SUMMARY_POS_PIPELINE = 0;
-    public static final int SUMMARY_POS_FEED = 1;
-    public static final int SUMMARY_POS_PRIORITY = 2;
-    public static final int SUMMARY_POS_STATUS = 3;
     public static final String FIELD_CREATE_TIME = "Created";
     public static final String FIELD_START_TIME = "Start Time";
     public static final String FIELD_END_TIME_DATE = "End Time";
-    public static final String FIELD_FEED_NAME = StreamDataSource.FEED_NAME;
+    public static final String FIELD_FEED = "Feed";
     public static final String FIELD_PRIORITY = "Priority";
-    public static final String FIELD_PIPELINE_UUID = StreamDataSource.PIPELINE_UUID;
+    public static final String FIELD_PIPELINE = "Pipeline";
     public static final String FIELD_STATUS = "Status";
     public static final String FIELD_COUNT = "Count";
     public static final String FIELD_NODE = "Node";